The Sekketian Shinobi (New Rare Adv Class) (#4072)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This adds the Sekketian Shinobi, a new "glass-cannon" like rare
adventurer class that is meant to fulfill a "ninja" niche and is meant
to reference a faction within the Queendom of Lakkari's lore.
<img width="60" height="97" alt="image"
src="https://github.com/user-attachments/assets/f508c9f5-407e-4f53-9beb-3c9d25f8559b"
/>

They have good combat skills, but lack any money or decent armor other
than a padded gambeson.


## Why It's Good For The Game

Ninjas are cool, and this also adds an additional lakkarian role to the
game., I saw some people in the past complain that there wasnt really
any lakkarian roles outisde the cleric - I wanted to change that. It
also isnt limited to just elves too, and is open to some more species.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Added a new role
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>