(BOUNTY) Collar of Servitude. Explosive collars to keep the adepts in line! (#4097)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
<img width="1836" height="339" alt="image"
src="https://github.com/user-attachments/assets/84778aa5-fdd3-4dda-9b72-4ace6e5fd60e"
/>

Every single adept now spawns with a collar of servitude, an explosive
collar that can be detonated or unlocked by a detonator that spawns with
the inquisitor. Ensuring their compliance to the inquisition.

Upon getting the detonation order from the detonator the collar plays a
short music box before promptly exploding
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

##

 Why It's Good For The Game
Video of testing:



https://github.com/user-attachments/assets/548a2787-f885-4e11-9189-fcd1a2be920c


I mean, just look at it. it's fucking rad yeah? oh and I guess it adds a
very good enforcement mechanism for the inquisitor for unruly adepts..
also this is a ook bounty.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: The Inquisition has fitted the every adepts with a 'Collar of
Servitude'. Ensuring their obedience and faithfulness to the
inquisition.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>