Cultivable mushrooms (#3771)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Trying this again. Huge thanks to Cheffie for all of the help.
### Ready for TM, but please do not full merge yet!
Two cultivars have been bred to aid in the growing of food and fibers in
areas which receive little of Astrata's light, capillus mort and
borowiki mushrooms. Three other wild types have begun to pop up in the
outlands of Vanderlin. Waddles are found exclusively in the murderwood,
caveweeps are found exclusively in Malum's Anvil, and merkels will spawn
very rarely in both and the bog.
This PR aims to expand the variety of cultivable crops within Vanderlin.
These crops will not require sunlight to grow.

- Adds a new family of crops to the game: Dikarya - Mushrooms. (yes,
that is actually a subkingdom, pretend it isn't.)
- Can be grown in soil as every other crop, but gets a boost to growth
and nutrient uptake if planted in mushroom mounds.
- Mushrooms have a special subtype of seeds called spores, acquired
exactly the same as normal crops.
- Adds a new soil subtype: the mushroom mound. Only mushrooms will be
able to grow well in the mushroom mound, other plants are not able to
form a proper root structure here, stunting them.
- Can be crafted with two nitesoil of any type and two chaff of any type
by attacking the nitesoil with the chaff.
- Behaves exactly like normal soil otherwise. Can be tilled, watered,
and fertilized as normal.
- Adds some mushroom spores and mushroom mounds to Vanderlin's dungeon
for the prisoners to grow.

The Agriculture book should be updated with all info pertaining to
these.

### Spriting work done by discord user noodle_buster. Check out their
other work!
The mushroom mound and its different states. Fully fertilized, watered,
and tilled down to unfertilized, unwatered, and untilled.
![ss+(2025-11-06+at+12 07
09)-1](https://github.com/user-attachments/assets/c222d385-b45b-4ec9-bde9-efb51e59a05c)

Current types of mushrooms and their uses:
Waddle: edible and tasty, can be added to stews or dried and made into a
filling tea / stock.
Capillus Mort: inedible (lightly toxic and gross) but can be dried to
produce fibers.
Merkel: edible, tasty, and a rare find, currently has no recipes.
Borowiki: edible and hearty, can be used for stew and a pie.
Caveweep: edibly and tasty, currently has no recipes.
<img width="346" height="66" alt="ss+(2025-11-06+at+12 04 42)-1"
src="https://github.com/user-attachments/assets/cc5162a7-df6f-4ae1-9f1b-6822f88853e9"
/>
![ss+(2025-11-05+at+11 51
21)](https://github.com/user-attachments/assets/716cbb18-cdc7-4252-a709-99cdd212e0af)
![ss+(2025-11-05+at+11 55
50)](https://github.com/user-attachments/assets/7dae9e17-3456-42cd-bac3-9a789b0005ae)
![ss+(2025-11-05+at+11 58
52)](https://github.com/user-attachments/assets/439eabdb-267b-4adf-b8bc-0901ff82234a)



The Vanderlin dungeon with the mushroom mounds and spores.
<img width="715" height="445" alt="ss+(2025-11-06+at+12 12 04)"
src="https://github.com/user-attachments/assets/e9fe769c-74a4-411e-956b-d6107fcfbdaf"
/>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
More crops and ways to grow them, as well as the ability to grow
substantial crops underground. Something meaningful for prisoners to do
and a proper way to learn farming while imprisoned.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
add: Wild and cultivable mushroom crops
add: Mushroom mound
map: Added mushroom mounds to Vanderlin's dungeon
map: Added debug mushroom mounds to Roguetest
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: die <96586172+grungussuss@users.noreply.github.com>