Wave Creation Tools for Admins Part 1 (#3937)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Thisi is a tool for admins to be used in events, they can create custom
jobs with custom outfits (or outfits existing), then create a wave with
custom or normal jobs to be spawned next to them.

This system is inspired in the CM ERTs.

This is the current part 1 of the update, i plan to do a part 2 with
more stuff (visible outfit on the outfit manager, accents tied to jobs
and etc)

- The tool is under debug and it is called "Wave Creation Tools", within
it you got 3 options on a pop up, Job Manage , Wave Manage and Outfit
Manage.

- The pop ups are per Admin but the Waves, Jobs and Outfits are global
and shared.

- Job Manage, with this tool you can create new jobs using almost all
options that a default job have, including antag, names, skills, stats,
traits, restrictions and others.
You can also Export, import and edit these jobs.
You can also equip someone with a custom job like a custom outfit
through the right click select equipment.
If you create a custom outfit for a job, it will be tied to the job so
when you export and eventually import it back the outfit will be added
as well.

- Wave Manage, with this tool you can create waves that function as ERT
to people that are either dead or at the lobby, these waves can have
Custom Jobs or just normal jobs, you can pick a job and the number of
slots, name for the wave, greet text and minimum of candidates.
It will fill the slots based on their order so 2 Servants, 2 Prince, if
there are 2 players it will fill the 2 Servant slots first.
If you put a minimum of X on a wave, unless there are X candidates
(people who pressed the join button) the wave won't start..
 You can also Export, Import and Edit these waves.
If you export a wave with 10 Custom jobs, when you import that same wave
again, it will add those custom jobs, so you don't need to manually
import different custom jobs. (This also includes these jobs unique
outfits)

- Outfit Manage, with this tool you can create new outfits from scratch,
it will pull from all the items that exist in the code and you can pick
from the list. (This one already exist but i am just improving it)
^Currently only the custom outfits are enabled until all the jobs
outfits are up to date with the new job stuff so they don't carry stat
and skills on these outfits.
You can also Export and Import these outfits.

- Important : The waves spawn under the admin and deploy after 60
seconds.

- [x] Create Custom Job
- [x] Delete Custom Job
- [x] Edit Custom Job
- [x] Import and Export Custom Job
- [x] Consistent browser even after refreshing for the lists (Job)
- [x] Consistent browser even after refreshing for the lists (Wave)
- [x] Create Wave
- [x] Delete Wave
- [x] Edit Wave
- [x] Improve Outfit Manager 
- [x] Admins can announce about calling  a wave
- [x] Wave spawning the mobs and their jobs correctly.

## Why It's Good For The Game

More tools

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
admin:  A tool to create custom waves for admins
admin: A tool to create custom jobs
admin: Improve Outfit Manager code make so it uses a list of items
instead of needing to type the path.
code: Move outfit code to wave manager
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>